refactor(clients): move to users routes
/users/:email/clients
