--- src/routes/clients.js
@@ -1,8 +1,0 @@
-const Router = require('koa-router');
-const ClientsController = require('../controllers/clients');
-
-const router = new Router();
-
-router.post('/register', ClientsController.register);
-
-module.exports = router;
